Improved version, hopefully more reliable
When running this forks version of ipwndfu, the device is in a state where we can read, write memory and execute code.
Furthermore wxn is disabled.
However to do sigpatches we still have the issue of ROM being not writable.

In my previous version i simply modified the pagetables and added write permissions to the existing rom mapping.
Then (as far as the MMU is concerned) the virtual addresses are writable.
But wait a sec, how is it possible to write to ROM, afterall it's ReadOnlyMemory?
Well, when the MMU is turned on, you also turn on cacheing, otherwise each memory access is painfully slow.
When writing to memory, you're also writing to CPU caches L1, L2, L3 and (depending on your cache policy) 
either directly to the main memory (write through policy), or the write is postponed until the cacheline is evicted from cache (write back policy).
So when mapping ROM writable and writing to it, we aren't actually writing to the ROM itself, but only to the CPU caches.

Now there are 2 issues we can run into with this method:
1) Assuming that performing an actual write request to ROM causes some kind of fault, the device might crash once that happens.
Assuming the cache is set up to be write-back (i didn't check what is actually used), then we're fine until the cacheline is evicted and an actual write to main
memory is performed. Now the ROM is mostly deterministic, but not fully deterministic. 
This means that if this is true, the method might work sometimes and sometimes we'll run into crashes (after extensively using the method, i indeed run into unrealiability issues).
If write-through policy is used, we would crash right away assuming a write to ROM does cause a crash.
Conclusion: We either have write-back cache policy, or a write to ROM does not cause a crash.

2) Even when a write to ROM does not cause a fault right away, we still have a problem in case the cachelines containing the patches are evicted from cache.
Because when the memory is fetched then, it will load the unmodified copy from ROM, rather our patched version.
Since the exploit does trash the heap, the device will panic when it reaches the heap corruption check before booting the next image.
Because there aren't any information about caching (how large the caches are etc), it's hard to tell which one of these is more likely the real issue.
However relying on patches to persist in CPU caches is non-ideal anyways.


There is a more obvious solution to the problem.
Rather than relying on the patches to stay alive in CPU caches, we can simply remap the pages in question and patch them in SRAM.
This is what i implemented here.
We need to apply 2 patches, which is: disabling the heap corruption check before jumping into the next image and disabling signature checks.
This will occupy 2 fully pages in SRAM (2*0x4000=0x8000) as well as some more space to set up another level of pagetables.
BootROM sets up just 2 Level page tables, but only uses the upper one to make block mappings of (if i'm not mistaken) 2MB.
One for the ROM, one for RAM and a couple others for device memory.
In order to do that we redicret the Level2 entry to a Level3 entry that we create and place in some unused memory, we only need to have enough page table entries to map the rom pages that are used, thus we don't need a full page here. It's fine if the page contains other data which are invalid page table entries, as long as BootROM doesn't access them.
The rest is easy: point L2 Table entry responsible for ROM to L3 Table which we create, in the L3 Table remap 2 pages to some unused pages at the end of SRAM.
Now since we have 2 mappings we don't even need to bother about wxn, since we can write to the SRAM by its *true* address and execute it with the ROM address.

This version should be more robust. Of course i only tested it twice, so i can't tell for sure. But i do believe it does :)
Be a believer too and merge this ;)